Move components under ui
