+++ src/client/ui/components/pluginlist.js
