+++ src/client/ui/components/settingsPanel.js
